Cache file is no longer required

+++ view-state.rb
@@ -179,16 +179,17 @@ $size_filename = "size.data"
 $size_cache_dirty = 0
 $size_cache = Hash.new(0)
 
-File.stat($size_filename).file? or usage('Size-file does not exist!', options)
-begin
-	File.open($size_filename, 'r').each_line { |line|
-		line.delete!( "\n" )
-		line.delete!( "\r" )
-		words = line.split( ':' )
-		$size_cache[ words[ 0 ] ] = words[ 1 ].to_i
-	}
-rescue
-	$stderr.print "Failure at #{$size_filename}: #{$!} => Skipping!\n"
+if ( FileTest.exist?($size_filename) )
+	begin
+		File.open($size_filename, 'r').each_line { |line|
+			line.delete!( "\n" )
+			line.delete!( "\r" )
+			words = line.split( ':' )
+			$size_cache[ words[ 0 ] ] = words[ 1 ].to_i
+		}
+	rescue
+		$stderr.print "Failure at #{$size_filename}: #{$!} => Skipping!\n"
+	end
 end
 
 $all_pack = 0
